requirements.txt
deta~=1.1.0
flask~=2.2.2
flask-restful
flask-cors
requests~=2.28.1
jsonpath~=0.82
pytz~=2022.6
Pillow~=9.3.0
DateTime~=4.7
Werkzeug~=2.2.2